pyproject.toml
[tool.poetry]
name = "fastapi-atom-chat"
version = "0.1.0"
description = "Test task for the Backend trainee developer vacancy in Rosatom"
authors = ["nickAnhel <anhimovn1@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.0"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
bcrypt = "^4.2.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
python-multipart = "^0.0.12"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
env_override_existing_values = 1
env_files = [
    ".test.env",
]
filterwarnings = [
    "ignore::FutureWarning",
]